// src/app/models/image-create-helper.ts
import { OponentEnum } from "../games/platform/models/oponent-enum";

export class ImageCreateHelper {
    src?: string;
    drawPointWidthMultiplier?: number;
    drawPointHeightMultiplier?: number;
    widthMultiplier?: number;
    heightMultiplier?: number;
    repetableBreakBetween?: number;
    isOnFloor?: boolean;
    oponentType?: OponentEnum;
}